fix: evita fallo de build sin variables de Supabase
