+++ src/app/dashboard/page.tsx
@@ -1,21 +1,5 @@
-'use client';
-
-import { useEffect } from 'react';
-import { useRouter } from 'next/navigation';
-import { supabase } from '@/lib/supabase';
+import DashboardPage from '@/components/DashboardPage';
 
 export default function Dashboard() {
-  const router = useRouter();
-
-  useEffect(() => {
-    supabase.auth.getSession().then(({ data }) => {
-      if (!data.session) router.replace('/');
-    });
-  }, [router]);
-
-  return (
-    <div className="min-h-screen flex items-center justify-center">
-      <h1 className="text-2xl text-white">Bienvenido al tablero</h1>
-    </div>
-  );
+  return <DashboardPage />;
 }
